Created viewModel for the Joke Fragment and created live data and observable for the Joke in fragment
