+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
 
     defaultConfig {
         applicationId "com.RaphaelAGN.chucknorrisapp"
-        minSdkVersion 19
+        minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
@@ -36,16 +36,37 @@ dependencies {
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     def fragment_version = "1.3.2"
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.3.2'
-    implementation "androidx.fragment:fragment-ktx:$fragment_version"
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.3.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    implementation 'com.squareup.retrofit2:retrofit:2.7.2'
-    implementation 'com.squareup.retrofit2:converter-gson:2.7.2'
-    implementation 'com.squareup.okhttp3:okhttp:3.14.7'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    dependencies {
+        def lifecycle_version = "2.3.0"
+        def arch_version = "2.1.0"
+
+        // ViewModel
+        implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
+        // LiveData
+        implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_version"
+        // Lifecycles only (without ViewModel or LiveData)
+        implementation "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycle_version"
+
+        // Saved state module for ViewModel
+        implementation "androidx.lifecycle:lifecycle-viewmodel-savedstate:$lifecycle_version"
+
+        // Jetpack Compose Integration for ViewModel
+        implementation "androidx.lifecycle:lifecycle-viewmodel-compose:1.0.0-alpha03"
+
+        // Annotation processor
+        implementation "androidx.lifecycle:lifecycle-common-java8:$lifecycle_version"
+
+        implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+        implementation 'androidx.core:core-ktx:1.3.2'
+        implementation "androidx.fragment:fragment-ktx:$fragment_version"
+        implementation 'androidx.appcompat:appcompat:1.2.0'
+        implementation 'com.google.android.material:material:1.3.0'
+        implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+        implementation 'com.squareup.retrofit2:retrofit:2.7.2'
+        implementation 'com.squareup.retrofit2:converter-gson:2.7.2'
+        implementation 'com.squareup.okhttp3:okhttp:3.14.7'
+        testImplementation 'junit:junit:4.13.2'
+        androidTestImplementation 'androidx.test.ext:junit:1.1.2'
+        androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    }
 }
