+++ app/src/main/java/com/RaphaelAGN/chucknorrisapp/fragments/MainFragment.kt
@@ -9,19 +9,19 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
-import android.widget.Toast
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.Observer
+import com.RaphaelAGN.chucknorrisapp.MainActivity
 import com.RaphaelAGN.chucknorrisapp.R
-import com.RaphaelAGN.chucknorrisapp.data.JokeModel
-import com.RaphaelAGN.chucknorrisapp.endpoint.Endpoint
-import com.RaphaelAGN.chucknorrisapp.retrofitClient.RetrofitClient
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+
+import com.RaphaelAGN.chucknorrisapp.viewModels.JokeViewModel
 
 class MainFragment : Fragment(R.layout.fragment_main) {
     var jokeButton: Button? = null
     var jokeTextView: TextView? = null
     val TAG = "chuck_norris_app"
+
+    val model: JokeViewModel by viewModels()
 
     val buttonText = "New Joke"
     val defaultTextOnTextView = "Click the button below to show a joke"
@@ -30,7 +30,9 @@ class MainFragment : Fragment(R.layout.fragment_main) {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
         Log.i(TAG, "onCreateViewFragment")
+
         val root = inflater.inflate(R.layout.fragment_main, container, false)
         jokeTextView = root.findViewById(R.id.joke_text_view)
 
@@ -40,28 +42,16 @@ class MainFragment : Fragment(R.layout.fragment_main) {
         jokeButton?.text = buttonText
 
         jokeButton?.setOnClickListener(){
-            getRandomJoke()
+            context?.let{
+                    context -> model.getRandomJoke(context)
+            }
         }
 
-        return root
-    }
-
-    fun getRandomJoke() {
-        val retrofitClient =
-                RetrofitClient.getRetrofitInstance("https://api.chucknorris.io/jokes/")
-
-        val endpoint = retrofitClient.create(Endpoint::class.java)
-        val callback = endpoint.getJoke()
-
-        callback.enqueue(object : Callback<JokeModel> {
-            override fun onFailure(call: Call<JokeModel>, t: Throwable) {
-                Toast.makeText(context, "An error has occurred", Toast.LENGTH_LONG).show()
-            }
-
-            override fun onResponse(call: Call<JokeModel>, response: Response<JokeModel>) {
-                jokeTextView?.text = response.body()?.value.toString()
-            }
+        model.currentJoke.observe(viewLifecycleOwner, Observer{
+            jokeTextView?.text = it
         })
+
+        return root
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
